% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMDCEV.R
\name{BayesMDCEV}
\alias{BayesMDCEV}
\title{BayesMDCEV}
\usage{
BayesMDCEV(stan_data, bayes_options, initial.parameters,
  keep.samples = FALSE, include.stanfit = TRUE)
}
\arguments{
\item{stan_data}{data for model}

\item{bayes_options}{list of Bayes options}

\item{initial.parameters}{Specify initial parameters intead of
starting at random.}

\item{keep.samples}{default is FALSE,}

\item{include.stanfit}{default isTRUE,}

\item{show.stan.warnings}{= TRUE}
}
\description{
Fit a MDCEV model using Bayesian estimation and Stan
}
