% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CleanInit}
\alias{CleanInit}
\title{CleanInit}
\usage{
CleanInit(init_input)
}
\arguments{
\item{init_input}{cleaned version of inits}
}
\description{
Adds appropriate dimensions for initial values
}
\keyword{internal}
