% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CombinePsiPhiVariables}
\alias{CombinePsiPhiVariables}
\title{GrabIndividualParms}
\usage{
CombinePsiPhiVariables(dat_id, dat, sim_rand)
}
\arguments{
\item{est_sim}{est_sim from results}

\item{parm_name}{name of parameter to get simulations}
}
\description{
Pulls out specific mdcev parameter simulations
}
\keyword{internal}
