% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateListsRow}
\alias{CreateListsRow}
\title{CreateListsRow}
\usage{
CreateListsRow(x)
}
\arguments{
\item{x}{matrix to be converted to list}
}
\value{
A list
}
\description{
Convert matrix x to a list with each row as an element
}
\examples{

tmp <- matrix(0, nrow = 10, ncol = 5)
tmp_list <- CreateListsRow(tmp)

}
