% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreatePsi}
\alias{CreatePsi}
\title{CreatePsi}
\usage{
CreatePsi(dat_vars_i, est_pars_i, J, NPsi_ij, psi_ascs, npols)
}
\arguments{
\item{dat_vars_i}{psi data for each person}

\item{est_pars_i}{psi parameter estimates for each person}
}
\description{
Works at individual level and creates the psi
variables for each simulation, policy, alternative
}
\keyword{internal}
