% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreatePsiMatrix}
\alias{CreatePsiMatrix}
\title{CreatePsiMatrix}
\usage{
CreatePsiMatrix(psi_j = NULL, psi_i = NULL)
}
\arguments{
\item{psi_j}{matrix (JXn_psi_j) of good-specific attributes}

\item{psi_i}{matrix (IXn_psi_i) of people-specific attributes}
}
\description{
Creates the Psi data matrix for use in mdcev model
}
\keyword{internal}
