% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-recreation.R
\docType{data}
\name{data_rec}
\alias{data_rec}
\title{Recreation data from Value of Nature to Canadians Survey}
\format{
A tibble with 16949 rows and 9 variables
}
\source{
\href{http://donnees.ec.gc.ca/data/managementoversight/systems/canadian-nature-survey-2012/}{Canadian Nature Survey 2012}
}
\usage{
data(data_rec)
}
\description{
Data from 997 individuals from the Value of Nature to Canadians (VNC) survey.
The travel costs are calculated using the approach described in Lloyd-Smith (2019)
}
\references{
Federal, Provincial, and Territorial Governments of Canada. 2014.
“2012 Canadian Nature Survey: Awareness, Participation, and Expenditures in Nature-Based Recreation, Conservation, and Subsistence Activities.” Ottawa, ON: Canadian Councils of Resource Ministers.
(\href{http://publications.gc.ca/site/eng/9.698872/publication.html}{PubMed})

Lloyd-Smith, P (2019).
“The Economic Benefits of Outdoor Recreation in Canada”. Working Paper
}
\keyword{datasets}
