% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateParmInfo.R
\name{GenClassNames}
\alias{GenClassNames}
\title{GenClassNames}
\usage{
GenClassNames(parms_names, n_classes)
}
\arguments{
\item{parms_names}{list of parameter names}

\item{n_classes}{The number of latent classes.}
}
\value{
A vector of LC names
}
\description{
Create names for LC
}
