% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{GrabIndividualParms}
\alias{GrabIndividualParms}
\title{GrabIndividualParms}
\usage{
GrabIndividualParms(est_sim, parm_name)
}
\arguments{
\item{est_sim}{est_sim from results}

\item{parm_name}{name of parameter to get simulations}
}
\description{
Pulls out specific mdcev parameter simulations
}
\keyword{internal}
