% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{GrabParms}
\alias{GrabParms}
\title{GrabParms}
\usage{
GrabParms(data, parm_name)
}
\arguments{
\item{data}{est_par object from results}

\item{parm_name}{name of parameter to get simulations}
}
\description{
Pulls out specific mdcev parameter simulations
}
