% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikeMDCEV.R
\name{maxlikeMDCEV}
\alias{maxlikeMDCEV}
\title{maxlikeMDCEV}
\usage{
maxlikeMDCEV(dat, n_classes = 1, n_draws = 50, seed = 123,
  initial.parameters = NULL, hessian = TRUE, keep_loglik)
}
\description{
Fit a MDCEV model MLE
}
