% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcev.data.R
\name{mdcev.data}
\alias{mdcev.data}
\alias{[[<-.mdcev.data}
\alias{$<-.mdcev.data}
\alias{print.pseries}
\alias{index}
\alias{mean.mdcev.data}
\alias{formula.mdcev.data}
\alias{print.mdcev.data}
\title{data.frame for mdcev model}
\usage{
mdcev.data(data, choice = NULL, alt.var = NULL, chid.var = NULL,
  alt.levels = NULL, id.var = NULL, group.var = NULL,
  drop.index = FALSE, subset = NULL, ...)
}
\arguments{
\item{data}{a `data.frame`,}

\item{choice}{the variable indicating the choice made: it has to be
a numerical vector}

\item{alt.var}{the name of the variable that contains the
alternative index or the name under which the alternative index
will be stored (the default  name is `alt`),}

\item{alt.levels}{the name of the alternatives: if null,
they are guessed from the `alt.var` argument,}

\item{id.var}{the name of the variable that contains the individual
index if any,}

\item{group.var}{the name of the variable that contains the group
index if any,}

\item{subset}{a logical expression which defines the subset of
observations to be selected,}

\item{x, object}{a `mdcev.data` or a `pseries` object,}
}
\value{
A `mdcev.data` object, which is a `data.frame` in `long`
    format, *i.e.* one line for each alternative.  It has a `index`
    attribute, which is a `data.frame` that contains the index of
    the choice made (`chid`), the index of the alternative (`alt`)
    and, if any, the index of the individual (`id`) and of the
    alternative groups (`group`).
}
\description{
shape a `data.frame` in a suitable form for the use of the
`mdcev` function and complete some data checks
}
