% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{MultiplyMatrix}
\alias{MultiplyMatrix}
\title{MultiplyMatrix}
\usage{
MultiplyMatrix(mat_temp, x, n_rows)
}
\arguments{
\item{mat_temp}{matrix A to be multipled}

\item{x}{matrix B to be multiplied}

\item{n_rows}{number of rows for final matrix}
}
\description{
Convert list to matrix by row
}
\keyword{internal}
