% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSimulationData.R
\name{PrepareSimulationData}
\alias{PrepareSimulationData}
\title{PrepareSimulationData}
\usage{
PrepareSimulationData(stan_est, policies, nsims = 30,
  price_change_only = FALSE)
}
\arguments{
\item{stan_est}{Stan fit model from FitMDCEV}

\item{policies}{list containing
price_p with additive price increases, and
dat_psi_p with new psi data}

\item{nsims}{Number of simulation draws to use for parameter uncertainty}

\item{price_change_only}{Choose if there is no change in dat_psi_p variables}
}
\value{
A list with individual-specific data (df_indiv) and common data (df_common)
and n_classes for number of classes and model_num for model type
}
\description{
Prepare Data for WTP simulation
}
