% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMDCEVdata.R
\name{processMDCEVdata}
\alias{processMDCEVdata}
\title{processMDCEVdata}
\usage{
processMDCEVdata(data, psi_formula, lc_formula, model_options)
}
\arguments{
\item{data}{The (IxJ) data to be passed to Stan including 1) id, 2) good, 3) quant,
4) price, 5) income, and columns for psi variables. Arrange data by id then good.
Notes I is number of individuals and J is number of non-numeraire goods.}

\item{psi_formula}{Formula for psi}

\item{lc_formula}{Formula for latent class}

\item{model_options}{list of model options}
}
\description{
Process MDCEV data
}
\keyword{internal}
