% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdcev-package.R
\docType{package}
\name{rmdcev}
\alias{rmdcev-package}
\alias{rmdcev}
\title{rmdcev: Estimating and simulating Kuhn-Tucker and multiple discrete-continuous extreme value (MDCEV) demand models}
\description{
The rmdcev R package estimate and simulates Kuhn-Tucker demand models with individual
heterogeneity. The models supported by rmdcev are the multiple-discrete continuous extreme value (MDCEV)
model and Kuhn-Tucker specification common in the environmental economics literature on recreation demand.
Latent class and random parameters specifications can be implemented and the models are fit using maximum
likelihood estimation or Bayesian estimation. All models are implemented in Stan, which is a C++ package
for performing full Bayesian inference (see Stan Development Team, 2019) <https://mc-stan.org/>. The rmdcev package also implements
demand forecasting (Pinjari and Bhat (2011) <https://repositories.lib.utexas.edu/handle/2152/23880>) and
welfare calculation (Lloyd-Smith (2018) <doi:10.1016/j.jocm.2017.12.002>) for policy simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plloydsmith/rmdcev}
  \item Report bugs at \url{https://github.com/plloydsmith/rmdcev/issues}
}

}
\author{
\strong{Maintainer}: Patrick Lloyd-Smith \email{patrick.lloydsmith@usask.ca}

Other contributors:
\itemize{
  \item Trustees of Columbia University [copyright holder]
}

}
\keyword{internal}
