% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateMdcevData.R
\name{SimulateMdcevData}
\alias{SimulateMdcevData}
\title{SimulateMdcevData}
\usage{
SimulateMdcevData(model, nobs = 1000, ngoods = 10, inc_lo = 1e+05,
  inc_hi = 50000, price_lo = 5, price_hi = 100, alpha_parms = 0.8,
  scale_parms = 1, gamma_parms = 1 + runif(ngoods, 0, 2),
  psi_i_parms = c(-1.5, 3, -2, 1, 2), psi_j_parms = c(-5, 0.5, 2),
  nerrs = 15)
}
\arguments{
\item{model}{A string indicating which model specification is estimated.
The options are "alpha","les", "gamma", and "gamma0".}

\item{nobs}{Number of individuals}

\item{ngoods}{Number of non-nuemraire goods}

\item{inc_lo}{Mean income}

\item{inc_hi}{Standard deviation of income}

\item{price_lo}{Mean price for non-numeraire goods}

\item{price_hi}{Standard deviation of price}

\item{alpha_parms}{Parameter value for alpha term}

\item{scale_parms}{Parameter value for scale term}

\item{gamma_parms}{Parameter value for gamma terms}

\item{psi_i_parms}{Parameter value for psi terms that vary by individual}

\item{psi_j_parms}{Parameter value for psi terms that vary by good}

\item{nerrs}{Number of error draws for demand simulation}
}
\value{
data for stan model
}
\description{
Simulate data for MDCEV model
}
