% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateWTP.R
\name{SimulateWTP}
\alias{SimulateWTP}
\title{SimulateWTP}
\usage{
SimulateWTP(stan_est, policies, nerrs = 30, nsims = 30,
  cond_error = 1, algo_gen = NULL)
}
\arguments{
\item{stan_est}{Stan fit model from FitMDCEV}

\item{policies}{list containing
price_p with additive price increases, and
dat_psi_p with new psi data}

\item{nerrs}{Number of error draws for welfare analysis}

\item{nsims}{Number of simulation draws to use for parameter uncertainty}

\item{cond_error}{Choose whether to draw errors conditional on actual demand or not.
Conditional error draws (=1) or unconditional error draws.}

\item{algo_gen}{Type of algorhitm for simulation. algo_gen = 0 for Hybrid Approach (i.e. constant alphas,
only model 3/4) alog_gen = 1 for General approach (i.e. heterogeneous alpha's, all models)}
}
\value{
wtp a list for each individual holding a nsims x npols matrix of wtp
}
\description{
Simulate WTP for MDCEV model
}
