% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateMDCEV.R
\name{SummaryDemand}
\alias{SummaryDemand}
\title{SummaryDemand}
\usage{
SummaryDemand(demand, ci = 0.95)
}
\arguments{
\item{demand}{list of welfare changes from SimulateDemand}

\item{ci}{confidence interval (for 95\% input 0.95)}
}
\value{
demand_sum summary table of demand results
}
\description{
Provide a summary of demands each policy
}
