% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMDCEV.R
\name{BayesMDCEV}
\alias{BayesMDCEV}
\title{BayesMDCEV}
\usage{
BayesMDCEV(
  stan_data,
  bayes_options,
  keep.samples = FALSE,
  include.stanfit = TRUE,
  ...
)
}
\arguments{
\item{stan_data}{data for model}

\item{bayes_options}{list of Bayes options}

\item{keep.samples}{default is FALSE,}

\item{include.stanfit}{default isTRUE,}

\item{...}{Additional parameters to pass on to \code{rstan::stan}
and \code{rstan::sampling}.}
}
\description{
Fit a MDCEV model using Bayesian estimation and Stan
}
\keyword{internal}
