% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMDCEV.R
\name{BayesMDCEV}
\alias{BayesMDCEV}
\title{BayesMDCEV}
\usage{
BayesMDCEV(stan_data, bayes_options, initial.parameters,
  keep.samples = FALSE, include.stanfit = TRUE)
}
\arguments{
\item{stan_data}{data for model}

\item{bayes_options}{list of Bayes options}

\item{initial.parameters}{Specify initial parameters intead of starting at random.
Initial parameter values should be included in a named list. For the "hybrid" specification,
initial parameters can be specified as:
init = list(psi = array(0, dim = c(1, num_psi)),
            gamma = array(1, dim = c(1, num_goods)),
            alpha = array(0.5, dim = c(1, 0)),
            scale = array(1, dim = c(1)))
where num_psi is number of psi parameters and num_goods is number of non-numeraire goods}

\item{keep.samples}{default is FALSE,}

\item{include.stanfit}{default isTRUE,}
}
\description{
Fit a MDCEV model using Bayesian estimation and Stan
}
\keyword{internal}
