% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMDCEV.R
\name{BayesMDCEV}
\alias{BayesMDCEV}
\title{BayesMDCEV}
\usage{
BayesMDCEV(
  stan_data,
  bayes_options,
  initial.parameters,
  keep.samples = FALSE,
  include.stanfit = TRUE,
  ...
)
}
\arguments{
\item{stan_data}{data for model}

\item{bayes_options}{list of Bayes options}

\item{initial.parameters}{The default for fixed and random parameter specifications is to use random starting values.
For LC models, the default is to use slightly adjusted MLE point estimates from the single class model.
Initial parameter values should be included in a named list. For the "hybrid" specification,
initial parameters can be specified as:
inititial.parameters = list(psi = array(runif(1), dim = c(K, num_psi)),
                            gamma = array(1, dim = c(K, num_alt)),
                            alpha = array(0.5, dim = c(K, 0)),
                            scale = array(1, dim = c(K)))
where K is the number of classes (i.e. K = 1 is used for single class models),
num_psi is number of psi parameters, and num_alt is number of non-numeraire alternatives.}

\item{keep.samples}{default is FALSE,}

\item{include.stanfit}{default isTRUE,}

\item{...}{Additional parameters to pass on to \code{rstan::stan}
and \code{rstan::sampling}.}
}
\description{
Fit a MDCEV model using Bayesian estimation and Stan
}
\keyword{internal}
