% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMdcevData.R
\name{CheckMdcevData}
\alias{CheckMdcevData}
\title{CheckMdcevData}
\usage{
CheckMdcevData(data)
}
\arguments{
\item{data}{The (IxJ) data to be passed to Stan including 1) id, 2) good, 3) quant,
4) price, 5) income, and columns for psi variables. Arrange data by id then good.
Notes I is number of individuals and J is number of non-numeraire goods.}
}
\description{
Check MDCEV data
}
