% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateBlankPolicies}
\alias{CreateBlankPolicies}
\title{CreateBlankPolicies}
\usage{
CreateBlankPolicies(npols, ngoods, dat_psi, price_change_only)
}
\arguments{
\item{npols}{Number of policies to simulate}

\item{ngoods}{Number of non-numeraire goods}

\item{dat_psi}{Psi data matrix used in estimation}

\item{price_change_only}{Logical value for whether to include policy changes to dat_psi.
TRUE implies that only price changes are used in simulation.}
}
\description{
Create 'zero effect' policies that can be modified
}
\examples{
\dontrun{
CreateBlankPolicies(npols = 2, ngoods = 10, dat_psi = NULL, price_change_only = TRUE)
}
}
