% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateBlankPolicies}
\alias{CreateBlankPolicies}
\title{CreateBlankPolicies}
\usage{
CreateBlankPolicies(npols, ngoods, dat_psi)
}
\description{
Create 'zero effect' policies that can be modified
}
