% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateBlankPolicies}
\alias{CreateBlankPolicies}
\title{CreateBlankPolicies}
\usage{
CreateBlankPolicies(npols, ngoods, dat_psi, price_change_only)
}
\arguments{
\item{npols}{Number of policies to simulate}

\item{ngoods}{Number of non-numeraire goods}

\item{dat_psi}{Psi data matrix used in estimation}
}
\description{
Create 'zero effect' policies that can be modified
}
