% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateBlankPolicies}
\alias{CreateBlankPolicies}
\title{CreateBlankPolicies}
\usage{
CreateBlankPolicies(npols, nalts, dat_psi, price_change_only = TRUE)
}
\arguments{
\item{npols}{Number of policies to simulate}

\item{nalts}{Number of non-numeraire alts}

\item{dat_psi}{Psi data matrix used in estimation}

\item{price_change_only}{Logical value for whether to include policy changes to dat_psi. Defaults to TRUE.
TRUE implies that only price changes are used in simulation.}
}
\description{
Create 'zero effect' policies that can be modified
}
\examples{
\donttest{
CreateBlankPolicies(npols = 2, nalts = 10, dat_psi = NULL, price_change_only = TRUE)
}
}
