% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{CreateListsCol}
\alias{CreateListsCol}
\title{CreateListsCol}
\usage{
CreateListsCol(x)
}
\arguments{
\item{x}{matrix to be converted to list}
}
\description{
Convert matrix x to a list with each row as an element
}
