% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateParmInfo.R
\name{CreateParmInfo}
\alias{CreateParmInfo}
\title{CreateParmInfo}
\usage{
CreateParmInfo(stan_data, algorithm, random_parameters)
}
\arguments{
\item{stan_data}{data for model}

\item{algorithm}{Either "Bayes" for Bayes or "MLE"
for maximum liklihood estimation.}

\item{random_parameters}{The form of the covariance matrix for
Bayes. Can be 'fixed', 'uncorr, 'corr'.}
}
\value{
A list of parameter names and numbers
}
\description{
Create parameter names and number of parameters
}
