% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-recreation.R
\docType{data}
\name{data_rec}
\alias{data_rec}
\title{Recreation data from Value of Nature to Canadians Survey}
\format{
A tibble with 34000 rows and 8 variables
}
\source{
\href{https://data-donnees.az.ec.gc.ca/data/managementoversight/systems/canadian-nature-survey-2012/}{Canadian Nature Survey 2012}
}
\usage{
data(data_rec)
}
\description{
Data from 2000 individuals from the Value of Nature to Canadians (VNC) survey.
The travel costs are calculated using the approach described in Lloyd-Smith (2020)
}
\references{
Federal, Provincial, and Territorial Governments of Canada. 2014.
“2012 Canadian Nature Survey: Awareness, Participation, and Expenditures in Nature-Based Recreation, Conservation, and Subsistence Activities.” Ottawa, ON: Canadian Councils of Resource Ministers.
(\href{https://publications.gc.ca/site/eng/9.698872/publication.html}{PubMed})

Lloyd-Smith, P (2022).
“The Economic Benefits of Recreation in Canada”. Canadian Journal of Economics.
\doi{10.1111/caje.12560}
}
\keyword{datasets}
