% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMDCEV.R
\name{FitMDCEV}
\alias{FitMDCEV}
\title{FitMDCEV}
\usage{
FitMDCEV(dat, dat_class = NULL, weights = NULL, price_num = NULL,
  model_specification = NULL, n_classes = 1, fixed_scale = 0,
  seed = 123, initial.parameters = NULL, algorithm = "HB-Stan",
  print_ll = 0, hessian = TRUE, n_draws = 50, keep_loglik = 0,
  hb_random_parameters = hb_random_parameters, n_iterations = 100,
  n_chains = 4, hb.max.tree.depth = 10, hb.adapt.delta = 0.8,
  hb.keep.samples = FALSE, hb.stanfit = TRUE, hb.lkj.prior.shape = 4,
  hb.warnings = TRUE, hb.beta.draws.to.keep = 0)
}
\arguments{
\item{weights}{An optional vector of sampling or frequency weights.}

\item{n_classes}{The number of latent classes.}

\item{seed}{Random seed.}

\item{initial.parameters}{Specify initial parameters intead of
starting at random in latent class analysis. The initial
parameters need to be supplied as list consisting of a matrix
called class.parameters whose columns are the parameters of the
classes, and a vector called class.sizes containing the class
size parameters.}

\item{algorithm}{Either "HB" for Hierarchical Bayes or "MLE"
for latent class analysis.}

\item{hb.max.tree.depth}{http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded}

\item{hb.adapt.delta}{http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup}

\item{hb.keep.samples}{Whether to keep the samples of all the
parameters in the output.}

\item{hb.stanfit}{Whether to include the stanfit property.}

\item{hb.lkj.prior.shape}{Real number greater than one; the shape
    hyperparameter for the LKJ prior used for the correlation matrix
    of the respondent coefficients distribution. A value of one gives
equal probability weight to all possible correlation matrices. Larger values
favour less correlation (draws closer to the identity matrix).}

\item{hb.warnings}{Whether to show warnings from Stan.}

\item{hb.beta.draws.to.keep}{Maximum number of beta draws per
respondent to return in beta.draws.}

\item{stan.dat}{The data to be passed to Stan.}

\item{normal.covariance}{The form of the covariance matrix for
Hierarchical Bayes. Can be 'Full, 'Spherical', 'Diagonal'.}

\item{hb.iterations}{The number of iterations in Hierarchical
Bayes.}

\item{hb.chains}{The number of chains in Hierarchical Bayes.}

\item{include.choice.parameters}{Whether to include
alternative-specific parameters.}

\item{...}{Additional parameters to pass on to \code{rstan::stan}
and \code{rstan::sampling}.}
}
\description{
Fit a MDCEV model using MLE or HB
}
