% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateParmInfo.R
\name{GenClassNames}
\alias{GenClassNames}
\title{GenClassNames}
\usage{
GenClassNames(parms_names, n_classes)
}
\arguments{
\item{parms_names}{list of parameter names}

\item{n_classes}{The number of latent classes. Note that the LC model is automatically estimated as long as the
prespecified number of classes is set greater than 1.}
}
\value{
A vector of LC names
}
\description{
Create names for LC
}
\keyword{internal}
