% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalBayesMDCEV.R
\name{HierarchicalBayesMDCEV}
\alias{HierarchicalBayesMDCEV}
\title{HierarchicalBayesMDCEV}
\usage{
HierarchicalBayesMDCEV(dat, model_type, n_iterations = 500,
  n_chains = 4, n_cores = 4, initial.parameters = initial.parameters,
  seed = 123, max.tree.depth = 10, adapt.delta = 0.8,
  keep.samples = FALSE, n_classes = 1, include.stanfit = TRUE,
  hb_random_parameters = "fixed", show.stan.warnings = TRUE,
  hb.lkj.prior.shape = hb.lkj.prior.shape)
}
\description{
Fit a MDCEV model HB
}
