% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikeMDCEV.R
\name{maxlikeMDCEV}
\alias{maxlikeMDCEV}
\title{maxlikeMDCEV}
\usage{
maxlikeMDCEV(stan_data, initial.parameters, mle_options, parms_info, ...)
}
\arguments{
\item{stan_data}{data for model fromatted from processMDCEVdata}

\item{initial.parameters}{The default for fixed and random parameter specifications is to use random starting values.
For LC models, the default is to use slightly adjusted MLE point estimates from the single class model.
Initial parameter values should be included in a named list. For example, the LC "hybrid" specification
initial parameters can be specified as:
initial.parameters = list(psi = array(0, dim = c(K, num_psi)),
                            gamma = array(1, dim = c(K, num_alt)),
                            alpha = array(0.5, dim = c(K, 0)),
                            scale = array(1, dim = c(K)))
where K is the number of classes (i.e. K = 1 is used for single class models),
num_psi is number of psi parameters, and num_alt is number of non-numeraire alternatives.}

\item{mle_options}{modeling options for MLE}

\item{parms_info}{information on parameters}

\item{...}{Additional parameters to pass on to \code{rstan::stan}
and \code{rstan::sampling}.}
}
\description{
Fit a MDCEV model with MLE
}
