% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikeMDCEV.R
\name{maxlikeMDCEV}
\alias{maxlikeMDCEV}
\title{maxlikeMDCEV}
\usage{
maxlikeMDCEV(stan_data, initial.parameters, mle_options)
}
\arguments{
\item{stan_data}{data for model fromatted from processMDCEVdata}

\item{initial.parameters}{Specify initial parameters intead of starting at random.
Initial parameter values should be included in a named list. For the "hybrid" specification,
initial parameters can be specified as:
init = list(psi = array(0, dim = c(1, num_psi)),
            gamma = array(1, dim = c(1, num_alt)),
            alpha = array(0.5, dim = c(1, 0)),
            scale = array(1, dim = c(1)))
where num_psi is number of psi parameters and num_alt is number of non-numeraire alternatives}

\item{mle_options}{modeling options for MLE}
}
\description{
Fit a MDCEV model with MLE
}
