% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikeMDCEV.R
\name{maxlikeMDCEV}
\alias{maxlikeMDCEV}
\title{maxlikeMDCEV}
\usage{
maxlikeMDCEV(stan_data, initial.parameters, seed, mle_options)
}
\arguments{
\item{stan_data}{data for model}

\item{initial.parameters}{Specify initial parameters intead of
starting at random.}

\item{seed}{Random seed.}

\item{mle_options}{modeling options for MLE}
}
\description{
Fit a MDCEV model with MLE
}
