% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMDCEVdata.R
\name{processMDCEVdata}
\alias{processMDCEVdata}
\title{processMDCEVdata}
\usage{
processMDCEVdata(data, psi_formula, lc_formula, num_price, model_options)
}
\arguments{
\item{data}{The (IxJ) data to be passed to Stan arranged by id then good including
1) quant2) price 3) income. Notes I is number of individuals and J is number of non-numeraire goods.}

\item{psi_formula}{Formula for psi}

\item{lc_formula}{Formula for latent class}

\item{num_price}{An optional vector containing price of numeraire or outside good (default is 1).}

\item{model_options}{list of model options}
}
\description{
Process MDCEV data
}
