% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMDCEVdata.R
\name{processMDCEVdata}
\alias{processMDCEVdata}
\title{processMDCEVdata}
\usage{
processMDCEVdata(formula, data, model_options)
}
\arguments{
\item{formula}{Formula for the model to be estimated.The formula is divided in
two parts, separated by the symbol \code{|}. The first part is reserved for
variables in the psi parameter. These can include alternative-specific and
individual-specific variables. The second part corresponds for individual-specific
variables that enter in the probability assignment in models with latent classes.}

\item{data}{The (IxJ) data to be passed to Stan of class \code{\link[rmdcev]{mdcev.data}}
including 1) id, 2) alt, 3) quant, 4) price, 5) income, and columns for psi variables.
Arrange data by id then alt. Note: I is number of individuals and J is number of
non-numeraire alternatives.}

\item{model_options}{list of model options}
}
\description{
Process MDCEV data
}
\keyword{internal}
