% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSimulationData.R
\name{ProcessSimulationData}
\alias{ProcessSimulationData}
\title{ProcessSimulationData}
\usage{
ProcessSimulationData(est_sim, stan_est, policies, nsims)
}
\arguments{
\item{est_sim}{Cleaned up parameter simulations from PrepareSimulationData}

\item{stan_est}{Stan fit model from FitMDCEV}

\item{policies}{list containing
price_p with additive price increases, and
dat_psi_p with new psi data}

\item{nsims}{Number of simulation draws to use for parameter uncertainty}
}
\description{
Internal function for WTP simulation
}
\keyword{internal}
