% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdcev-package.R
\docType{package}
\name{rmdcev}
\alias{rmdcev}
\alias{rmdcev-package}
\title{The 'rmdcev' package.}
\description{
The rmdcev R package estimates and simulates multiple discrete-continuous extreme value (MDCEV)
demand models with observed and unobserved individual heterogneity. Fixed parameter, latent class,
and random parameter models can be estimated. These models are estimated using maximum likelihood or
Bayesian estimation techniques and are implemented in Stan, which is a C++ package for performing full
Bayesian inference (see http://mc-stan.org/). The package also supports Phaneuf and von Haefen's Kuhn-Tucker
model specification.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
