% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMDCEV.R
\name{RunStanSampling}
\alias{RunStanSampling}
\title{RunStanSampling}
\usage{
RunStanSampling(stan_data, stan.model, bayes_options)
}
\arguments{
\item{stan_data}{data for model}

\item{stan.model}{Complied Stan model}

\item{bayes_options}{list of Bayes options}

\item{...}{Additional parameters to pass on to \code{rstan::stan} and
\code{rstan::sampling}.}
}
\value{
A stanfit object.
}
\description{
Wrapper function for \code{rstan:stan} and
\code{rstan:sampling} to run Stan Bayes analysis.
}
\keyword{internal}
