% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateWTP.R
\name{SimulateWTP}
\alias{SimulateWTP}
\title{SimulateWTP}
\usage{
SimulateWTP(df_indiv, df_common, nerrs = 30, cond_error = 1,
  algo_gen = NULL)
}
\arguments{
\item{df_indiv}{Prepared individual level data from PrepareSimulationData}

\item{df_common}{Prepared common data from PrepareSimulationData}

\item{nerrs}{Number of error draws for welfare analysis}

\item{cond_error}{Choose whether to draw errors conditional on actual demand or not.
Conditional error draws (=1) or unconditional error draws.}

\item{algo_gen}{Type of algorhitm for simulation. algo_gen = 0 for Hybrid Approach (i.e. constant alphas,
only model 3/4) alog_gen = 1 for General approach (i.e. heterogeneous alpha's, all models)}
}
\value{
wtp a list for each individual holding a nsims x npols matrix of wtp
}
\description{
Simulate WTP for MDCEV model
}
