% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanSimulate.R
\name{StanDemand}
\alias{StanDemand}
\title{StanDemand}
\usage{
StanDemand(df_indiv, df_common, sim_options)
}
\arguments{
\item{df_indiv}{list of income, quant_j, price, psi, and psi_p that vary by individual}

\item{df_common}{list of parameters that are constant for all individuals}

\item{sim_options}{list of simualtion options}
}
\value{
demand with nsim lists of npolsXngoods+1 matrices
}
\description{
Use Stan functions to simulate Marshallian demand
}
\keyword{internal}
