% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryMDCEV.R
\name{SummaryMDCEV}
\alias{SummaryMDCEV}
\title{SummaryMDCEV}
\usage{
SummaryMDCEV(model, printCI = FALSE)
}
\arguments{
\item{model}{Model object returned by function \link{FitMDCEV}.}

\item{printCI}{set to TRUE to print 95\% confidence intervals}
}
\value{
A matrix of coefficients, s.d. and z-tests (invisible)
}
\description{
Prints mdcev estimation results to console. The output format is borrowed from the Apollo package
}
