% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateWTP.R
\name{SummaryWelfare}
\alias{SummaryWelfare}
\title{SummaryWelfare}
\usage{
SummaryWelfare(wtp, ci = 0.95)
}
\arguments{
\item{wtp}{list of welfare changes from SimulateWTP}

\item{ci}{confidence interval (for 95\% input 0.95)}
}
\value{
wtp_sum summary table of welfare results
}
\description{
Provide a summary of welfare changes for each policy
}
